refactor pipeline: rename scripts for clarity in `src` directory and update corresponding references in `README.md` and RPV notebook
