+++ filtering.py
@@ -11,7 +11,7 @@ from pathlib import Path
 def main():
     """Entry point for the uav-filter command."""
     # Import the main function from the script file
-    script_path = Path(__file__).parent / "02_filtering.py"
+    script_path = Path(__file__).parent / "pipeline_filtering.py"
     spec = importlib.util.spec_from_file_location("filtering_script", script_path)
     module = importlib.util.module_from_spec(spec)
     sys.modules["filtering_script"] = module
