+++ rpv_modelling.py
@@ -11,7 +11,7 @@ from pathlib import Path
 def main():
     """Entry point for the uav-rpv command."""
     # Import the main function from the script file
-    script_path = Path(__file__).parent / "03_RPV_modelling.py"
+    script_path = Path(__file__).parent / "pipeline_RPV_modelling.py"
     spec = importlib.util.spec_from_file_location("rpv_modelling_script", script_path)
     module = importlib.util.module_from_spec(spec)
     sys.modules["rpv_modelling_script"] = module
