--- src/Util/__init__.py
@@ -1,1 +1,0 @@
-# Package marker for Util
