+++ src/Utils/extract_data/polygon_filtering.py
