+++ src/pipeline_filtering.py
@@ -1,20 +1,16 @@
 import argparse
 import glob
-import logging
 import os
 
 import polars as pl
 
 from src.Common.config_object import config_object
-from src.Common.data_loader import load_by_polygon
-from src.Common.filters import (
+from src.Utils.filtering.data_loader import load_by_polygon
+from src.Utils.filtering.filters import (
     OSAVI_index_filtering,
-    add_mask_and_plot,
     excess_green_filter,
-    plot_heatmap,
-    plot_spectrogram,
 )
-from src.Util.logging import logging_config
+from src.Common.logging import logging_config
 
 
 def main():
